resultContent.tsxに分けた
